refactored chat drawer file
