+++ resources/views/livewire/chat/chat.blade.php
@@ -116,5 +116,5 @@
 
     </div>
 
-    <livewire:wirechat.chat-drawer />
+    <livewire:wirechat.chat.drawer />
 </div>
